VKT(Backend): OPHVKTKEH-185 moved scheduled actions code under ClerkEnrollmentService

+++ backend/vkt/src/main/java/fi/oph/vkt/service/ClerkEnrollmentService.java
@@ -19,9 +19,13 @@ import fi.oph.vkt.util.ClerkEnrollmentUtil;
 import fi.oph.vkt.util.UUIDSource;
 import fi.oph.vkt.util.exception.APIException;
 import fi.oph.vkt.util.exception.APIExceptionType;
+import java.time.Duration;
 import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
@@ -30,6 +34,8 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class ClerkEnrollmentService extends AbstractEnrollmentService {
+
+  private static final Logger LOG = LoggerFactory.getLogger(ClerkEnrollmentService.class);
 
   private final EnrollmentRepository enrollmentRepository;
   private final ExamEventRepository examEventRepository;
@@ -105,22 +111,36 @@ public class ClerkEnrollmentService extends AbstractEnrollmentService {
   }
 
   @Transactional(isolation = Isolation.SERIALIZABLE)
-  public void cancelUnfinishedEnrollment(final long enrollmentId) {
-    final Enrollment enrollment = enrollmentRepository.getReferenceById(enrollmentId);
-    assert enrollment.isUnfinished();
+  public void cancelPassiveEnrollmentsExpectingPayment() {
+    final Duration paymentTime = Duration.of(3, ChronoUnit.HOURS);
 
-    enrollment.setStatus(EnrollmentStatus.CANCELED_UNFINISHED_ENROLLMENT);
-    enrollmentRepository.saveAndFlush(enrollment);
+    enrollmentRepository
+      .findAllByStatus(EnrollmentStatus.EXPECTING_PAYMENT_UNFINISHED_ENROLLMENT)
+      .stream()
+      .filter(e -> e.getModifiedAt().plus(paymentTime).isBefore(LocalDateTime.now()))
+      .forEach(enrollment -> {
+        enrollment.setStatus(EnrollmentStatus.CANCELED_UNFINISHED_ENROLLMENT);
+        enrollmentRepository.saveAndFlush(enrollment);
+      });
   }
 
   @Transactional(isolation = Isolation.SERIALIZABLE)
-  public void deleteEnrollment(final long enrollmentId) {
-    final Enrollment enrollment = enrollmentRepository.getReferenceById(enrollmentId);
-    final List<Payment> payments = enrollment.getPayments();
+  public void deleteCanceledUnfinishedEnrollments() {
+    final Duration ttl = Duration.of(24, ChronoUnit.HOURS);
 
-    assert payments.stream().noneMatch(p -> p.getPaymentStatus() == PaymentStatus.OK);
+    enrollmentRepository
+      .findAllByStatus(EnrollmentStatus.CANCELED_UNFINISHED_ENROLLMENT)
+      .stream()
+      .filter(e -> e.getModifiedAt().plus(ttl).isBefore(LocalDateTime.now()))
+      .forEach(enrollment -> {
+        final List<Payment> payments = enrollment.getPayments();
 
-    paymentRepository.deleteAllInBatch(enrollment.getPayments());
-    enrollmentRepository.deleteById(enrollment.getId());
+        if (payments.stream().anyMatch(p -> p.getPaymentStatus() == PaymentStatus.OK)) {
+          LOG.warn(String.format("Tried to delete enrollment (%d) with paid payment", enrollment.getId()));
+        } else {
+          paymentRepository.deleteAllInBatch(payments);
+          enrollmentRepository.deleteById(enrollment.getId());
+        }
+      });
   }
 }
